Fix static init order issue
