+++ table/block_based/filter_policy.cc
@@ -1310,72 +1310,6 @@ Status XXPH3FilterBitsBuilder::MaybePostVerify(const Slice& filter_content) {
 }
 }  // namespace
 
-// For testing only
-namespace test {
-
-class LegacyBloomFilterPolicy : public BloomLikeFilterPolicy {
- public:
-  explicit LegacyBloomFilterPolicy(double bits_per_key)
-      : BloomLikeFilterPolicy(bits_per_key) {}
-
-  FilterBitsBuilder* GetBuilderWithContext(
-      const FilterBuildingContext& context) const override {
-    if (GetMillibitsPerKey() == 0) {
-      // "No filter" special case
-      return nullptr;
-    }
-    return GetLegacyBloomBuilderWithContext(context);
-  }
-
-  static constexpr const char* kName = "rocksdb.internal.LegacyBloomFilter";
-  std::string ToString() const override {
-    return kName + GetBitsPerKeySuffix();
-  }
-};
-
-class FastLocalBloomFilterPolicy : public BloomLikeFilterPolicy {
- public:
-  explicit FastLocalBloomFilterPolicy(double bits_per_key)
-      : BloomLikeFilterPolicy(bits_per_key) {}
-
-  FilterBitsBuilder* GetBuilderWithContext(
-      const FilterBuildingContext& context) const override {
-    if (GetMillibitsPerKey() == 0) {
-      // "No filter" special case
-      return nullptr;
-    }
-    return GetFastLocalBloomBuilderWithContext(context);
-  }
-
-  static constexpr const char* kName = "rocksdb.internal.FastLocalBloomFilter";
-  std::string ToString() const override {
-    return kName + GetBitsPerKeySuffix();
-  }
-};
-
-class Standard128RibbonFilterPolicy : public BloomLikeFilterPolicy {
- public:
-  explicit Standard128RibbonFilterPolicy(double bloom_equiv_bits_per_key)
-      : BloomLikeFilterPolicy(bloom_equiv_bits_per_key) {}
-
-  FilterBitsBuilder* GetBuilderWithContext(
-      const FilterBuildingContext& context) const override {
-    if (GetMillibitsPerKey() == 0) {
-      // "No filter" special case
-      return nullptr;
-    }
-    return GetStandard128RibbonBuilderWithContext(context);
-  }
-
-  static constexpr const char* kName =
-      "rocksdb.internal.Standard128RibbonFilter";
-  std::string ToString() const override {
-    return kName + GetBitsPerKeySuffix();
-  }
-};
-
-}  // namespace test
-
 BloomLikeFilterPolicy::BloomLikeFilterPolicy(double bits_per_key)
     : warned_(false), aggregate_rounding_balance_(0) {
   // Sanitize bits_per_key
@@ -1593,6 +1527,38 @@ FilterBitsBuilder* BuiltinFilterPolicy::GetBuilderFromContext(
     return nullptr;
   }
 }
+
+// For testing only, but always constructable with internal names
+namespace test {
+
+FilterBitsBuilder* LegacyBloomFilterPolicy::GetBuilderWithContext(
+    const FilterBuildingContext& context) const {
+  if (GetMillibitsPerKey() == 0) {
+    // "No filter" special case
+    return nullptr;
+  }
+  return GetLegacyBloomBuilderWithContext(context);
+}
+
+FilterBitsBuilder* FastLocalBloomFilterPolicy::GetBuilderWithContext(
+    const FilterBuildingContext& context) const {
+  if (GetMillibitsPerKey() == 0) {
+    // "No filter" special case
+    return nullptr;
+  }
+  return GetFastLocalBloomBuilderWithContext(context);
+}
+
+FilterBitsBuilder* Standard128RibbonFilterPolicy::GetBuilderWithContext(
+    const FilterBuildingContext& context) const {
+  if (GetMillibitsPerKey() == 0) {
+    // "No filter" special case
+    return nullptr;
+  }
+  return GetStandard128RibbonBuilderWithContext(context);
+}
+
+}  // namespace test
 
 BuiltinFilterBitsReader* BuiltinFilterPolicy::GetBuiltinFilterBitsReader(
     const Slice& contents) {
